[DPE-2613] fix rel tests (#263)
## Issue
After updating to Juju 3.1 the charm failed on its relation tests #247

This was due to when apps scale down in 3.1, they receive relation
broken events. Thereby attempts to remove users were made when relations
were actually not removed, but instead a unit was being removed.

## Solution
Add a check in `relation-departed`. In this function we can tell if we
are receiving the `relation-broken` event due to scale down or relation
removal.

## more
Once this PR is accepted, we will make another one for branch `5/edge`

---------

Co-authored-by: Mykola Marzhan <303592+delgod@users.noreply.github.com>