[DPE-3292] test mongos relation (#341)
## Issue
Validation of mongos/cluster relation doesn't exist

## Solution
Test that relation does what its supposed to do i.e. user creation /
removal

## Files to not review
Feel free to ignore the test charm we copied over to
`tests/integration/sharding_tests/application/`, this should be removed
in the futre

## Future work
Ideally we can deploy a "Test Charm" for mongos to be hosted on instead
of this current charm that we have copied over
`tests/integration/sharding_tests/application/`. We can either:
1. test with `data-integrator` when it works with `mongos`
2. publish a [test-charm](https://charmhub.io/mysql-test-app) and deploy
it in the tests