[DPE-3832] Fix incorrect statuses for Adding shards (#383)
## Issue
#380 - Shards incorrectly report `Adding shard to config-server` when
they cannot reach the config-server. This may occur for three reasons:
1. The shard is truly getting added to the config-server and it does not
have access to the config-server
2. The shard is getting its cluster password updated
3. The shard is getting updating TLS

In cases 2&3, the shard is not getting added to the cluster and so this
status is wrong

## Solution
Report that the shard is getting added to the cluster only when it is
truly getting added for the first time.

Reporting this status in `relation-changed` does not make sense since
this hook is used to manage changes in the sharding interface. Therefore
it makes the most sense to report this when the shard is actually
getting added to the cluster (i.e. relation-joined)