fix (tests): ensure that the leader is managing secrets (#356)
## Issue

As of ops 2.10+ `Harness` applies the same access controls to secrets as
Juju does in production. This causes tests to fail that were ignoring
these access restrictions (deliberately - they have docstrings that call
it out).

## Solution

Set the unit to be leader in each of the applicable tests. It might be
that you'd rather have tests for both cases, leader and (failing
appropriately) non-leader - I can adjust to do that if you prefer.

---------

Co-authored-by: Mehdi Bendriss <bendrissmehdi@gmail.com>