[DPE-4494] Revision checking (#419)
## Issue
Currently Charmed MongoDB supports creating clusters with different
versions of charm components, i.e. a shard with revision 88 and a config
server with revision 100. this can cause unexpected issues.

## Solution
Prevent the charm from doing this by adding a revision check on relation
joined and reporting inconsistencies in update status

## Future PRs
To keep the size of this PR down we will do integration and unit tests
in the subsequent PRs
If Pedro and other team member wish to use it we can create this as a
lib in the data platform repo

---------

Co-authored-by: Mehdi Bendriss <bendrissmehdi@gmail.com>