[DPE-4143] Add a post upgrade check for cluster upgrades (#429)
## Issue
After the cluster upgrades the balancer stays disabled (a requirement
for the upgrade procedure to happen)

## Solution
Wait for all cluster components to successfully upgrade and then verify
cluster health. Once this is done we:
1. Re-enable the balancer
2. Set the new feature version across the cluster

## Other issues found along the way
The check for status across the cluster is faulty in these ways:
1. If the upgrade fails and sets the unit status to being
Failed/Blocked, it prevents the post upgrade check from ever succeeding
2. It is not flexible in how it checks mismatched cluster statuses
across the cluster.

Charm has no idea when a local cluster has completed an upgrade.
(Solution was to add an internal charm revision)

## Charm internal version CI workflow
This workflow ensures that we bump the number on this file each time we
make a PR.
We expect this workflow to fail right now since the file does not exist
on main branch. For that reason we expect it to fail. Once the PR merges
we will make a requirement of all future PRs

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-19-96.eu-central-1.compute.internal>
Co-authored-by: Mehdi Bendriss <bendrissmehdi@gmail.com>
Co-authored-by: phvalguima <pedro.guimaraes@canonical.com>