fix: Temporary incorrect shard status (#501)
## Issue

During an update-status event, the shard status would show an incorrect
status "Charm is in sharding role: database. Does not support shard
interface".

## Solution

This is due to the newly added checks a few weeks ago on the update
relation app databag. Fixed by running this handler only if we're not a
shard.
Also, the log line was incorrect, two arguments were swapped.