patch: remove the auto-delete config option (#575)
## Issue
The `auto-delete` config option is no longer a desired config option in
release 8
The terraform checks throw and ambiguous relation error.

## Solution
This PR removes the `auto-delete` config option and its logic.
Now, the databases are never dropped.
This prevents unintended data loss in case of misconfiguration.
Adds the endpoint to the certificates relation in the TF modules